edit: Add basic support for input line editing
This allows for simple line editing with history capabilities. On
purpose this has no concept of terminal input or terminal output and
just allows manipulation of an internal wide character string.

The history storing and loading is currently missing, but would be easy
to add. The other two items are missing completion and hints handling.

This feature is not yet complete, but good enough for a first review and
a sample application using it in a curses environment will follow.

The debug option is something that might need to be removed or at least
changed a little bit, but right now it is nice to see the internal
states.