edit: Add basic support for input line editing
This allows for simple line editing with history capabilities. On
purpose this has no concept of terminal input or terminal output and
just allows manipulation of an internal wide character string.

The debug option is something that might need to be removed or at least
changed a little bit, but right now it is nice to see the internal
states.

Following features are missing:

1) Work with words (delete, move etc.)
2) Tab completion
3) Hints system

Following features are left to the user:

1) Showing the prompt
2) Switching to masked input

The demo-edit is just for demonstration purposes and requires Curses to
be available without any autoconf magic.

With demo-cli there is a really simple command line handling example
that uses an internal terminal abstraction.