netlink: Add l_netlink_iter
Move netlink attribute iterator out of genl.c and into netlink.c.  Both
genl and netlink use the same attribute structure, and is applicable to
both.  This makes implementing netlink and rtnl message parsers much
easier.

In order to keep source compatibility, structure the netlink iterator in
the same way as the genl iterator.  In the future, the use of genl
iterators will be removed.